Refactoring out all custom hooks into their own file and importing them to index.js for exports
